+++ yanguo200/project/code/Camera/Target.h
@@ -6,7 +6,7 @@
 #define Target_EER 		2
 
 void Find_Target1(void);
-void Find_Target2(int p1,int p2);
+void Find_Target2(int16 p1,int16 p2);
 void LowerCameraExposure(void);
 
 #endif
